add bsa support
- refactor main program and commands
